+++ packages/compiler-ssr/src/transforms/ssrVIf.ts
@@ -4,10 +4,7 @@ import {
   IfNode,
   createIfStatement,
   createBlockStatement,
-  createCallExpression,
-  IfBranchNode,
-  BlockStatement,
-  NodeTypes
+  createCallExpression
 } from '@vue/compiler-dom'
 import {
   SSRTransformContext,
@@ -26,14 +23,17 @@ export function ssrProcessIf(node: IfNode, context: SSRTransformContext) {
   const [rootBranch] = node.branches
   const ifStatement = createIfStatement(
     rootBranch.condition!,
-    processIfBranch(rootBranch, context)
+    processChildrenAsStatement(rootBranch.children, context)
   )
   context.pushStatement(ifStatement)
 
   let currentIf = ifStatement
   for (let i = 1; i < node.branches.length; i++) {
     const branch = node.branches[i]
-    const branchBlockStatement = processIfBranch(branch, context)
+    const branchBlockStatement = processChildrenAsStatement(
+      branch.children,
+      context
+    )
     if (branch.condition) {
       // else-if
       currentIf = currentIf.alternate = createIfStatement(
@@ -51,16 +51,4 @@ export function ssrProcessIf(node: IfNode, context: SSRTransformContext) {
       createCallExpression(`_push`, ['`<!---->`'])
     ])
   }
-}
-
-function processIfBranch(
-  branch: IfBranchNode,
-  context: SSRTransformContext
-): BlockStatement {
-  const { children } = branch
-  const needFragmentWrapper =
-    (children.length !== 1 || children[0].type !== NodeTypes.ELEMENT) &&
-    // optimize away nested fragments when the only child is a ForNode
-    !(children.length === 1 && children[0].type === NodeTypes.FOR)
-  return processChildrenAsStatement(children, context, needFragmentWrapper)
 }